verify password (#22)

+++ twimi-app/src/infrastructure/repositories/users.rs
@@ -37,7 +37,9 @@ impl UsersRepository for UsersRepositoryImpl {
         .await?
         .last_insert_id();
 
-        Ok(kaiin.try_into().map(|user: User| user.set_id(kaiin_id))?)
+        Ok(kaiin
+            .try_into()
+            .map(|user: User| user.set_id(kaiin_id as i64))?)
     }
 
     async fn find_user_by_username(
